Move LLS mixin to existing render package mixin
